moving graphql schema away from handler.js in order to guarantee testability.
